release/camera/requirements.txt
imageio==2.34.1
numpy==2.0.0
pillow==10.3.0
pyvirtualcam==0.11.1
